# api/openapi.yaml
openapi: 3.0.3
info:
  title: Karvon Scraper API
  version: 1.0.0
  description: |
    REST API for Karvon: the scraper module (Google Maps lead scraping with website
    email crawling) and the campaign module (outreach campaigns, contacts and their
    lifecycle, reusable email content, consent and the two delivery providers). Live
    job progress is delivered over Server-Sent Events.

    All routes except `/healthz`, `/openapi.json` and `/webhooks/*` require
    `Authorization: Bearer <KARVON_API_KEY>`. The `/webhooks/*` routes carry no API
    key: they are called by Instantly and Mailchimp, and are authenticated instead by
    the per-registration token in their path together with the secret that provider
    signs its deliveries with.
servers:
  - url: /api/v1
    description: Default mount point
security:
  - bearerAuth: []
...
paths:
...
  /stats/scraper:
    get:
      tags: [stats]
      operationId: getScraperStats
      summary: Dashboard counters and per-job email chart data
      responses:
        '200':
          description: Scraper statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScraperStats' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /sources:
    get:
      tags: [sources]
      operationId: listSources
      summary: List configured providers
      responses:
        '200':
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Source' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /sources/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [sources]
      operationId: getSource
      summary: Get a single provider
      responses:
        '200':
          description: Provider
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [sources]
      operationId: updateSource
      summary: Update provider key, cost and enabled flag
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SourceUpdate' }
      responses:
        '200':
          description: Updated provider
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /sources/{id}/test:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [sources]
      operationId: testSource
      summary: Validate the stored API key with a single-result provider call
      responses:
        '200':
          description: Key works
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceTestResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '502':
          description: Provider rejected the key or is unreachable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /newsletter/audiences:
    get:
      tags: [newsletter]
      operationId: listNewsletterAudiences
      summary: The Mailchimp audiences known here
      responses:
        '200':
          description: Audiences
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NewsletterAudience' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /newsletter/audiences/sync:
    post:
      tags: [newsletter]
      operationId: syncNewsletterAudiences
      summary: Queue a refresh of the audience list from Mailchimp
      responses:
        '202': { description: Sync queued }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /newsletter/audiences/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [newsletter]
      operationId: updateNewsletterAudience
      summary: Set the audience's opt-in policy, default tags and default flag
      description: |
        `allow_single_opt_in` decides whether a push asks Mailchimp for `subscribed` or
        for `pending`. Leaving it off means every contact still confirms by email,
        whatever consent we already hold.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewsletterAudienceUpdate' }
      responses:
        '200':
          description: Updated audience
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewsletterAudience' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /newsletter/audiences/{id}/webhook:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [newsletter]
      operationId: registerNewsletterWebhook
      summary: Register this audience's Mailchimp webhook
      responses:
        '200':
          description: Updated audience
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewsletterAudience' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502':
          description: Mailchimp refused the registration or is unreachable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [newsletter]
      operationId: deleteNewsletterWebhook
      summary: Remove this audience's Mailchimp webhook
      responses:
        '200':
          description: Updated audience
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewsletterAudience' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /newsletter/eligibility:
    get:
      tags: [newsletter]
      operationId: listNewsletterEligibility
      summary: Who could be pushed to the newsletter now, and who is one step short
      description: |
        Every row carries the consent gate's decision, including the reason a contact
        is not eligible, so the UI never has to re-derive the rule.
      parameters:
        - name: campaign_id
          in: query
          schema: { type: string, format: uuid }
        - name: audience_id
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated candidates
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EligibleContactList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /newsletter/push:
    post:
      tags: [newsletter]
      operationId: pushNewsletterContacts
      summary: Queue contacts for Mailchimp
      description: |
        The consent gate is applied per contact rather than to the request as a whole:
        contacts that pass are queued, contacts that do not come back in `rejected`
        with the reason, and the request still succeeds.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewsletterPushRequest' }
      responses:
        '200':
          description: What was queued and what was refused
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewsletterPushResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /newsletter/subscriptions:
    get:
      tags: [newsletter]
      operationId: listNewsletterSubscriptions
      summary: Every subscription with what we asked for and what Mailchimp says
      parameters:
        - name: status
          in: query
          description: Repeatable filter on Mailchimp's status
          schema:
            type: array
            items: { $ref: '#/components/schemas/SubscriptionStatus' }
          style: form
          explode: true
        - name: sync_status
          in: query
          description: Repeatable filter on our own push state
          schema:
            type: array
            items: { $ref: '#/components/schemas/SyncStatus' }
          style: form
          explode: true
        - name: audience_id
          in: query
          schema: { type: string, format: uuid }
        - name: q
          in: query
          description: Case-insensitive match on the contact's address
          schema: { type: string, maxLength: 200 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated subscriptions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewsletterSubscriptionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /newsletter/subscriptions/{id}/retry:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [newsletter]
      operationId: retryNewsletterSubscription
      summary: Re-queue a subscription whose push to Mailchimp failed
      responses:
        '200':
          description: The re-queued subscription
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewsletterSubscription' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /newsletter/stats:
    get:
      tags: [newsletter]
      operationId: getNewsletterStats
      summary: Subscription counts by status and the eligible → subscribed conversion
      responses:
        '200':
          description: Newsletter statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewsletterStats' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Static API key, `Authorization: Bearer <KARVON_API_KEY>`"
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Action not allowed in the resource's current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Request body or query failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string }
            details:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
      additionalProperties: false
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    PageMeta:
      type: object
      required: [page, per_page, total]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer, format: int64 }
...
    SourceKind:
      type: string
      enum: [apify, outscraper, emailable]
    SourceRole:
      type: string
      description: |
        What a source is for. Maps providers supply business listings; the verifier
        checks email addresses. Both store their key the same way.
      enum: [maps, verifier]
    Source:
      type: object
      required: [id, kind, role, name, has_key, cost_per_1k_cents, enabled, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/SourceKind' }
        role: { $ref: '#/components/schemas/SourceRole' }
        name: { type: string }
        has_key: { type: boolean, description: Whether an API key is stored; the key itself is never returned }
        cost_per_1k_cents: { type: integer, minimum: 0 }
        enabled: { type: boolean }
        last_tested_at: { type: string, format: date-time, nullable: true }
        last_test_ok: { type: boolean, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SourceUpdate:
      type: object
      properties:
        api_key:
          type: string
          nullable: true
          maxLength: 500
          description: Plaintext key; stored AES-GCM encrypted. Omit to keep, null to clear.
        cost_per_1k_cents: { type: integer, minimum: 0, maximum: 1000000 }
        enabled: { type: boolean }
        name: { type: string, minLength: 1, maxLength: 100 }
      additionalProperties: false
    SourceTestResult:
      type: object
      required: [ok, kind]
      properties:
        ok: { type: boolean }
        kind: { $ref: '#/components/schemas/SourceKind' }
        listings_returned: { type: integer, description: Maps providers only }
        credits:
          type: integer
          format: int64
          nullable: true
          description: Remaining balance, returned when testing the email verifier
        tested_at: { type: string, format: date-time }
    JobStatus:
      type: string
      enum: [queued, running, done, failed, cancelled]
...
    Job:
      type: object
      required: [id, name, status, source_id, config, stats, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { $ref: '#/components/schemas/JobStatus' }
        source_id: { type: string, format: uuid }
        source_name: { type: string }
        source_kind: { $ref: '#/components/schemas/SourceKind' }
        config: { $ref: '#/components/schemas/JobConfig' }
        stats: { $ref: '#/components/schemas/JobStats' }
        error: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
...
    ScraperStatsJobEmails:
      type: object
      required: [job_id, name, emails]
      properties:
        job_id: { type: string, format: uuid }
        name: { type: string }
        emails: { type: integer }
    ScraperStats:
      type: object
      required: [businesses, with_email, jobs_total, emails_per_job]
      properties:
        businesses: { type: integer, format: int64 }
        with_email: { type: integer, format: int64 }
        jobs_total: { type: integer, format: int64 }
        emails_total: { type: integer, format: int64 }
        last_job:
          allOf:
            - $ref: '#/components/schemas/Job'
          nullable: true
        emails_per_job:
          type: array
          items: { $ref: '#/components/schemas/ScraperStatsJobEmails' }
...
    SubscriptionStatus:
      type: string
      description: |
        What Mailchimp says about a subscription. `local_pending` is ours: the row
        exists here but has not been pushed yet.
      enum: [local_pending, pending, subscribed, unsubscribed, cleaned, transactional,
             archived, compliance_blocked, error]
    RequestedSubscriptionStatus:
      type: string
      description: What we asked Mailchimp for, which the audience's opt-in policy decides.
      enum: [pending, subscribed, unsubscribed]
    SyncStatus:
      type: string
      description: How far a subscription has got in being pushed to Mailchimp.
      enum: [queued, syncing, synced, failed]
...
    CountsByKey:
      type: object
      description: Counts keyed by a status, stage or other discriminator.
      additionalProperties: { type: integer, format: int64 }
...
    NewsletterAudience:
      type: object
      description: |
        A Mailchimp audience. `webhook_token` and the signing secret are never
        returned; `webhook_url` already carries what the client needs to show.
      required: [id, mailchimp_list_id, name, double_optin, allow_single_opt_in,
                 default_tags, is_default, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        mailchimp_list_id: { type: string }
        name: { type: string }
        double_optin: { type: boolean, description: What Mailchimp itself has configured }
        allow_single_opt_in:
          type: boolean
          description: |
            Whether a push may ask for `subscribed` directly. Off means every contact
            still confirms by email, whatever consent we hold.
        default_tags:
          type: array
          items: { type: string }
        webhook_id: { type: string, nullable: true }
        webhook_url: { type: string, nullable: true }
        member_count: { type: integer, nullable: true }
        stats:
          type: object
          additionalProperties: true
          nullable: true
        is_default: { type: boolean }
        last_synced_at: { type: string, format: date-time, nullable: true }
        last_sync_error: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NewsletterAudienceUpdate:
      type: object
      properties:
        allow_single_opt_in: { type: boolean }
        default_tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 60 }
        is_default: { type: boolean }
      additionalProperties: false
...
    EligibleContact:
      type: object
      required: [contact, decision]
      properties:
        contact: { $ref: '#/components/schemas/EligibleContactRecord' }
        decision: { $ref: '#/components/schemas/ConsentDecision' }
    EligibleContactList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/EligibleContact' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    NewsletterPushRequest:
      type: object
      required: [contact_ids]
      properties:
        contact_ids:
          type: array
          minItems: 1
          maxItems: 5000
          items: { type: string, format: uuid }
        audience_id:
          type: string
          format: uuid
          description: Defaults to the audience flagged as default
      additionalProperties: false
    NewsletterPushRejection:
      type: object
      required: [contact_id, reason]
      properties:
        contact_id: { type: string, format: uuid }
        reason: { type: string, example: no_consent }
    NewsletterPushResult:
      type: object
      description: |
        The gate is applied per contact, not to the request: what passes is queued and
        what does not comes back with its reason, rather than failing the whole push.
      required: [queued, rejected]
      properties:
        queued:
          type: array
          items: { type: string, format: uuid }
        rejected:
          type: array
          items: { $ref: '#/components/schemas/NewsletterPushRejection' }
    NewsletterSubscription:
      type: object
      required: [id, contact_id, audience_id, requested_status, status, sync_status,
                 sync_attempts, tags, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        contact_id: { type: string, format: uuid }
        audience_id: { type: string, format: uuid }
        consent_id: { type: string, format: uuid, nullable: true }
        requested_status: { $ref: '#/components/schemas/RequestedSubscriptionStatus' }
        status: { $ref: '#/components/schemas/SubscriptionStatus' }
        sync_status: { $ref: '#/components/schemas/SyncStatus' }
        subscriber_hash: { type: string, nullable: true }
        unique_email_id: { type: string, nullable: true }
        mailchimp_contact_id: { type: string, nullable: true }
        web_id: { type: integer, format: int64, nullable: true }
        claimed_at: { type: string, format: date-time, nullable: true }
        pushed_at: { type: string, format: date-time, nullable: true }
        last_synced_at: { type: string, format: date-time, nullable: true }
        last_error: { type: string, nullable: true }
        sync_attempts: { type: integer }
        subscribed_at: { type: string, format: date-time, nullable: true }
        unsubscribed_at: { type: string, format: date-time, nullable: true }
        unsubscribe_reason: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NewsletterSubscriptionList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/NewsletterSubscription' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    NewsletterStats:
      type: object
      required: [by_status, eligible, subscribed, pending, conversion_rate]
      properties:
        by_status: { $ref: '#/components/schemas/CountsByKey' }
        eligible: { type: integer, format: int64 }
        subscribed: { type: integer, format: int64 }
        pending: { type: integer, format: int64 }
        conversion_rate:
          type: number
          format: double
          description: Subscribers over eligible contacts
...